Move pandas related functionalities into pandas package
