+++ python/pyspark/sql/session.py
@@ -15,6 +15,8 @@
 # limitations under the License.
 #
 
+# To disallow implicit relative import. Remove this once we drop Python 2.
+from __future__ import absolute_import
 from __future__ import print_function
 import sys
 import warnings
@@ -25,15 +27,16 @@ if sys.version >= '3':
     basestring = unicode = str
     xrange = range
 else:
-    from itertools import izip as zip, imap as map
+    from itertools import imap as map
 
 from pyspark import since
 from pyspark.rdd import RDD, ignore_unicode_prefix
 from pyspark.sql.conf import RuntimeConfig
 from pyspark.sql.dataframe import DataFrame
+from pyspark.sql.pandas.conversion import SparkConversionMixin
 from pyspark.sql.readwriter import DataFrameReader
 from pyspark.sql.streaming import DataStreamReader
-from pyspark.sql.types import Row, DataType, StringType, StructType, TimestampType, \
+from pyspark.sql.types import Row, DataType, StringType, StructType, \
     _make_type_verifier, _infer_schema, _has_nulltype, _merge_type, _create_converter, \
     _parse_datatype_string
 from pyspark.sql.utils import install_exception_handler
@@ -60,7 +63,7 @@ def _monkey_patch_RDD(sparkSession):
     RDD.toDF = toDF
 
 
-class SparkSession(object):
+class SparkSession(SparkConversionMixin):
     """The entry point to programming Spark with the Dataset and DataFrame API.
 
     A SparkSession can be used create :class:`DataFrame`, register :class:`DataFrame` as
@@ -458,135 +461,6 @@ class SparkSession(object):
         data = [schema.toInternal(row) for row in data]
         return self._sc.parallelize(data), schema
 
-    def _get_numpy_record_dtype(self, rec):
-        """
-        Used when converting a pandas.DataFrame to Spark using to_records(), this will correct
-        the dtypes of fields in a record so they can be properly loaded into Spark.
-        :param rec: a numpy record to check field dtypes
-        :return corrected dtype for a numpy.record or None if no correction needed
-        """
-        import numpy as np
-        cur_dtypes = rec.dtype
-        col_names = cur_dtypes.names
-        record_type_list = []
-        has_rec_fix = False
-        for i in xrange(len(cur_dtypes)):
-            curr_type = cur_dtypes[i]
-            # If type is a datetime64 timestamp, convert to microseconds
-            # NOTE: if dtype is datetime[ns] then np.record.tolist() will output values as longs,
-            # conversion from [us] or lower will lead to py datetime objects, see SPARK-22417
-            if curr_type == np.dtype('datetime64[ns]'):
-                curr_type = 'datetime64[us]'
-                has_rec_fix = True
-            record_type_list.append((str(col_names[i]), curr_type))
-        return np.dtype(record_type_list) if has_rec_fix else None
-
-    def _convert_from_pandas(self, pdf, schema, timezone):
-        """
-         Convert a pandas.DataFrame to list of records that can be used to make a DataFrame
-         :return list of records
-        """
-        if timezone is not None:
-            from pyspark.sql.types import _check_series_convert_timestamps_tz_local
-            copied = False
-            if isinstance(schema, StructType):
-                for field in schema:
-                    # TODO: handle nested timestamps, such as ArrayType(TimestampType())?
-                    if isinstance(field.dataType, TimestampType):
-                        s = _check_series_convert_timestamps_tz_local(pdf[field.name], timezone)
-                        if s is not pdf[field.name]:
-                            if not copied:
-                                # Copy once if the series is modified to prevent the original
-                                # Pandas DataFrame from being updated
-                                pdf = pdf.copy()
-                                copied = True
-                            pdf[field.name] = s
-            else:
-                for column, series in pdf.iteritems():
-                    s = _check_series_convert_timestamps_tz_local(series, timezone)
-                    if s is not series:
-                        if not copied:
-                            # Copy once if the series is modified to prevent the original
-                            # Pandas DataFrame from being updated
-                            pdf = pdf.copy()
-                            copied = True
-                        pdf[column] = s
-
-        # Convert pandas.DataFrame to list of numpy records
-        np_records = pdf.to_records(index=False)
-
-        # Check if any columns need to be fixed for Spark to infer properly
-        if len(np_records) > 0:
-            record_dtype = self._get_numpy_record_dtype(np_records[0])
-            if record_dtype is not None:
-                return [r.astype(record_dtype).tolist() for r in np_records]
-
-        # Convert list of numpy records to python lists
-        return [r.tolist() for r in np_records]
-
-    def _create_from_pandas_with_arrow(self, pdf, schema, timezone):
-        """
-        Create a DataFrame from a given pandas.DataFrame by slicing it into partitions, converting
-        to Arrow data, then sending to the JVM to parallelize. If a schema is passed in, the
-        data types will be used to coerce the data in Pandas to Arrow conversion.
-        """
-        from pyspark.serializers import ArrowStreamPandasSerializer
-        from pyspark.sql.types import from_arrow_type, to_arrow_type, TimestampType
-        from pyspark.sql.utils import require_minimum_pandas_version, \
-            require_minimum_pyarrow_version
-
-        require_minimum_pandas_version()
-        require_minimum_pyarrow_version()
-
-        from pandas.api.types import is_datetime64_dtype, is_datetime64tz_dtype
-        import pyarrow as pa
-
-        # Create the Spark schema from list of names passed in with Arrow types
-        if isinstance(schema, (list, tuple)):
-            arrow_schema = pa.Schema.from_pandas(pdf, preserve_index=False)
-            struct = StructType()
-            for name, field in zip(schema, arrow_schema):
-                struct.add(name, from_arrow_type(field.type), nullable=field.nullable)
-            schema = struct
-
-        # Determine arrow types to coerce data when creating batches
-        if isinstance(schema, StructType):
-            arrow_types = [to_arrow_type(f.dataType) for f in schema.fields]
-        elif isinstance(schema, DataType):
-            raise ValueError("Single data type %s is not supported with Arrow" % str(schema))
-        else:
-            # Any timestamps must be coerced to be compatible with Spark
-            arrow_types = [to_arrow_type(TimestampType())
-                           if is_datetime64_dtype(t) or is_datetime64tz_dtype(t) else None
-                           for t in pdf.dtypes]
-
-        # Slice the DataFrame to be batched
-        step = -(-len(pdf) // self.sparkContext.defaultParallelism)  # round int up
-        pdf_slices = (pdf[start:start + step] for start in xrange(0, len(pdf), step))
-
-        # Create list of Arrow (columns, type) for serializer dump_stream
-        arrow_data = [[(c, t) for (_, c), t in zip(pdf_slice.iteritems(), arrow_types)]
-                      for pdf_slice in pdf_slices]
-
-        jsqlContext = self._wrapped._jsqlContext
-
-        safecheck = self._wrapped._conf.arrowSafeTypeConversion()
-        col_by_name = True  # col by name only applies to StructType columns, can't happen here
-        ser = ArrowStreamPandasSerializer(timezone, safecheck, col_by_name)
-
-        def reader_func(temp_filename):
-            return self._jvm.PythonSQLUtils.readArrowStreamFromFile(jsqlContext, temp_filename)
-
-        def create_RDD_server():
-            return self._jvm.ArrowRDDServer(jsqlContext)
-
-        # Create Spark DataFrame from Arrow stream file, using one batch per partition
-        jrdd = self._sc._serialize_to_jvm(arrow_data, ser, reader_func, create_RDD_server)
-        jdf = self._jvm.PythonSQLUtils.toDataFrame(jrdd, schema.json(), jsqlContext)
-        df = DataFrame(jdf, self._wrapped)
-        df._schema = schema
-        return df
-
     @staticmethod
     def _create_shell_session():
         """
@@ -722,46 +596,12 @@ class SparkSession(object):
         except Exception:
             has_pandas = False
         if has_pandas and isinstance(data, pandas.DataFrame):
-            from pyspark.sql.utils import require_minimum_pandas_version
-            require_minimum_pandas_version()
+            # Create a DataFrame from pandas DataFrame.
+            return super(SparkSession, self).createDataFrame(
+                data, schema, verifySchema, samplingRatio)
+        return self._create_dataframe(data, schema, verifySchema, samplingRatio)
 
-            if self._wrapped._conf.pandasRespectSessionTimeZone():
-                timezone = self._wrapped._conf.sessionLocalTimeZone()
-            else:
-                timezone = None
-
-            # If no schema supplied by user then get the names of columns only
-            if schema is None:
-                schema = [str(x) if not isinstance(x, basestring) else
-                          (x.encode('utf-8') if not isinstance(x, str) else x)
-                          for x in data.columns]
-
-            if self._wrapped._conf.arrowPySparkEnabled() and len(data) > 0:
-                try:
-                    return self._create_from_pandas_with_arrow(data, schema, timezone)
-                except Exception as e:
-                    from pyspark.util import _exception_message
-
-                    if self._wrapped._conf.arrowPySparkFallbackEnabled():
-                        msg = (
-                            "createDataFrame attempted Arrow optimization because "
-                            "'spark.sql.execution.arrow.pyspark.enabled' is set to true; however, "
-                            "failed by the reason below:\n  %s\n"
-                            "Attempting non-optimization as "
-                            "'spark.sql.execution.arrow.pyspark.fallback.enabled' is set to "
-                            "true." % _exception_message(e))
-                        warnings.warn(msg)
-                    else:
-                        msg = (
-                            "createDataFrame attempted Arrow optimization because "
-                            "'spark.sql.execution.arrow.pyspark.enabled' is set to true, but has "
-                            "reached the error below and will not continue because automatic "
-                            "fallback with 'spark.sql.execution.arrow.pyspark.fallback.enabled' "
-                            "has been set to false.\n  %s" % _exception_message(e))
-                        warnings.warn(msg)
-                        raise
-            data = self._convert_from_pandas(data, schema, timezone)
-
+    def _create_dataframe(self, data, schema, verifySchema, samplingRatio):
         if isinstance(schema, StructType):
             verify_func = _make_type_verifier(schema) if verifySchema else lambda _: True
